Password Generator
This program is built to generate the desired length of password given by the user. In this program, I imported the package random and string that helps to create a random password that consist strings, numbers, punctuation, lowercase and uppercase, etc. The program is able to generate the password of a fixed length given by the user.
How this program can be used ? 
1) It can be used for the strong suggestion of the passwords.
2) It helps the users to secure the online media handles with random generation of the password.
3) Bruteforcing is difficult to perform on the generated passwords.