Contact Book
You can store your contacts by the help of this program. It also helps you to search, delete, view the contacts by their name. This program can be used for storing the contacts in an efficient way. This program is designed in such an environment that makes it to response faster to the user about the operations.
You can have a look on the code uploaded here.